Add files via upload
The scope of the meat quality assessment dataset is to provide a comprehensive collection of data related to the physical and chemical properties of different types of meat, as well as subjective evaluations of meat quality. The dataset includes information on factors such as pH, color, tenderness, juiciness, and flavor, which are critical components of meat quality. It’s objective is to provide researchers and industry professionals with a tool to better understand the factors that contribute to meat quality and to develop strategies to improve meat quality and consumer satisfaction.
Overall, the meat quality assessment dataset provides a valuable resource for advancing research and innovation in the meat industry, ultimately leading to improved meat quality and increased consumer satisfaction.